+++ environment.h
@@ -27,14 +27,14 @@ public:
     void knockout_genes(const ops disabled[], int disabled_len);
 	void update_leadership();
 	void clear_last_moves();
-	float landmarkcoord[NUM_TESTS][2];
-	float hyenacoord[NUM_TESTS][NUM_HYENAS][2];
-	float lioncoord[NUM_TESTS][NUM_LIONS][2];
+	double landmarkcoord[NUM_TESTS][2];
+	double hyenacoord[NUM_TESTS][NUM_HYENAS][2];
+	double lioncoord[NUM_TESTS][NUM_LIONS][2];
 	int leadership[NUM_HYENAS];
 private:
 	team *agents;
-	float landmarkx;
-	float landmarky;
+	double landmarkx;
+	double landmarky;
 	QSet<int> unmoved;
 	QSet<int> uncalled;
 	QSet<int> just_moved;
